requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
environs==9.5.0
importlib-resources==5.7.1
inflection==0.5.1
jsonschema==4.4.0
marshmallow==3.15.0
packaging==21.3
Pillow==9.1.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
zipp==3.8.0
